Move show funcs to show.py, add categories opt to show_colors
